added unit tests for lotto hit by the player class

+++ src/main/java/pl/minigames/lotto/LottoHitByThePlayer.java
@@ -14,6 +14,6 @@ public class LottoHitByThePlayer {
 
         if (result) size = hitedNumbers.size();
 
-        return size;
+        return (Integer) size;
     }
 }
